Refactor Intro page to new file
