+++ package.json
@@ -38,7 +38,6 @@
   },
   "husky": {
     "hooks": {
-      "pre-push": "git branch | grep \"*\" | egrep -v \"^* master$\""
     }
   }
 }
