+++ src/api/README.md
@@ -1,0 +1,11 @@
+To set up your API key,
+
+- Get API key from https://openweathermap.org/
+- Create a file called api.js in the api folder
+- In the api.js file, paste the following lines of code:
+
+const key = 'YOUR_API_KEY_HERE'
+
+module.exports = key;
+
+- Make sure that your API key is between the ''
